Reorganize shared components and containers
