+++ app/routes/select-destination/index.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import AssignmentPanel from "../../shared-containers/AssignmentPanel"
+import AssignmentPanel from "../shared/containers/AssignmentPanel"
 import Footer from "./components/Footer"
 import ConfirmPanel from "./components/ConfirmPanel"
