+++ app/routes/select-submissions/components/Submission.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes } from "react"
-import ItemPanel from "../../../shared-components/ItemPanel"
+import ItemPanel from "../../shared/components/ItemPanel"
 
 const Submission = ({
   id,
