+++ app/routes/shared/components/ItemPanel.jsx
